+++ test/test.m
@@ -10,3 +10,4 @@ runtests('test_SNIRF.m')
 runtests('test_lufr.m')
 runtests('test_layout_merge.m')
 
+