+++ test/test_layout_merge.m
@@ -54,7 +54,14 @@ end
 
 
 %% Test 3: merge layout files
-lumofile.merge_layout(fullfile(lmpath, 'samples', 'sample_v040_no_layout.LUMO'));
+try
+  lumofile.merge_layout(fullfile(lmpath, 'samples', 'sample_v040_no_layout.LUMO'));
+  error('Merge layout should throw');
+catch e
+end
+  
+
+
 
 lumo_merged_fn = lumofile.merge_layout(fullfile(lmpath, 'samples', 'sample_v040_no_layout.LUMO'), ...
                                        'C:\Users\Sam\AppData\Local\Gowerlabs\Lumo\coordinates_268.json');